fix:email copy 로직 수정 / refactor: footer 컴포넌트 구조 분할
