+++ src/components/Footer/SocialLinks.js
@@ -1,0 +1,7 @@
+export const SOCIALLINKS = {
+  instagram: 'https://www.instagram.com/mju_likelion/',
+  github: 'https://github.com/mju-likelion',
+  facebook: 'https://www.facebook.com/likelionatmju',
+  medium: 'https://medium.com/@mju-likelion',
+  email: 'mju@likelion.org',
+};
